Dotty variants
