+++ test/shared/src/main/scala-dotty/zio/test/SourceLocationVariants.scala
@@ -1,0 +1,5 @@
+package zio.test
+
+trait SourceLocationVariants {
+  inline given SourceLocation = ${Macros.sourceLocation_impl}
+}
